Update D&D API endpoints and schema definitions
- Changed the API endpoint for fetching monsters to include the year in the URL.
- Updated the URL format for monster links in the list response.
- Modified the `challenge_rating` field type from Float to Integer in both the Marshmallow schema and the database model for consistency.

+++ backend/contributors/wainermora/dnd/schemas.py
@@ -52,7 +52,7 @@ class MonsterSchema(Schema):
     condition_immunities = fields.Raw(allow_none=True)
     senses = fields.Raw(allow_none=True)
     languages = fields.String(allow_none=True)
-    challenge_rating = fields.Float(allow_none=True)
+    challenge_rating = fields.Integer(allow_none=True)
     proficiency_bonus = fields.Integer(allow_none=True)
     xp = fields.Integer(allow_none=True)
     special_abilities = fields.Raw(allow_none=True)
